Correção de bugs

+++ app/src/main/java/com/glauco/weatherapp/db/fb/FBDatabase.kt
@@ -1,10 +1,10 @@
 package com.glauco.weatherapp.db.fb
 
-import com.google.firebase.Firebase
-import com.google.firebase.auth.auth
+import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.DocumentChange
 import com.google.firebase.firestore.ListenerRegistration
-import com.google.firebase.firestore.firestore
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
 
 class FBDatabase {
     interface Listener {
@@ -27,25 +27,23 @@ class FBDatabase {
                 listener?.onUserSignOut()
                 return@addAuthStateListener
             }
-
             val refCurrUser = db.collection("users").document(auth.currentUser!!.uid)
             refCurrUser.get().addOnSuccessListener {
                 it.toObject(FBUser::class.java)?.let { user ->
                     listener?.onUserLoaded(user)
                 }
             }
-
             citiesListReg = refCurrUser.collection("cities")
                 .addSnapshotListener { snapshots, ex ->
                     if (ex != null) return@addSnapshotListener
-
                     snapshots?.documentChanges?.forEach { change ->
                         val fbCity = change.document.toObject(FBCity::class.java)
-                        when (change.type) {
-                            DocumentChange.Type.ADDED -> listener?.onCityAdded(fbCity)
-                            DocumentChange.Type.MODIFIED -> listener?.onCityUpdated(fbCity)
-                            DocumentChange.Type.REMOVED -> listener?.onCityRemoved(fbCity)
-                            else -> {}
+                        if (change.type == DocumentChange.Type.ADDED) {
+                            listener?.onCityAdded(fbCity)
+                        } else if (change.type == DocumentChange.Type.MODIFIED) {
+                            listener?.onCityUpdated(fbCity)
+                        } else if (change.type == DocumentChange.Type.REMOVED) {
+                            listener?.onCityRemoved(fbCity)
                         }
                     }
                 }
@@ -57,27 +55,22 @@ class FBDatabase {
     }
 
     fun register(user: FBUser) {
-        if (auth.currentUser == null)
-            throw RuntimeException("User not logged in!")
+        if (auth.currentUser == null) throw RuntimeException("User not logged in!")
         val uid = auth.currentUser!!.uid
         db.collection("users").document(uid).set(user)
     }
 
     fun add(city: FBCity) {
-        if (auth.currentUser == null)
-            throw RuntimeException("User not logged in!")
-        if (city.name == null || city.name!!.isEmpty())
-            throw RuntimeException("City with null or empty name!")
+        if (auth.currentUser == null) throw RuntimeException("User not logged in!")
+        if (city.name == null || city.name!!.isEmpty()) throw RuntimeException("City with null or empty name!")
         val uid = auth.currentUser!!.uid
         db.collection("users").document(uid).collection("cities")
             .document(city.name!!).set(city)
     }
 
     fun remove(city: FBCity) {
-        if (auth.currentUser == null)
-            throw RuntimeException("User not logged in!")
-        if (city.name == null || city.name!!.isEmpty())
-            throw RuntimeException("City with null or empty name!")
+        if (auth.currentUser == null) throw RuntimeException("User not logged in!")
+        if (city.name == null || city.name!!.isEmpty()) throw RuntimeException("City with null or empty name!")
         val uid = auth.currentUser!!.uid
         db.collection("users").document(uid).collection("cities")
             .document(city.name!!).delete()
